+++ zetaclient/chains/evm/observer/v2_inbound.go
@@ -116,21 +116,21 @@ func (ob *Observer) observeGatewayDeposit(
 
 // parseAndValidateDepositEvents collects and sorts events by block number, tx index, and log index
 func (ob *Observer) parseAndValidateDepositEvents(
-	rawLogs []ethtypes.Log,
+	ethlogs []ethtypes.Log,
 	gatewayAddr ethcommon.Address,
 	gatewayContract *gatewayevm.GatewayEVM,
 ) []*gatewayevm.GatewayEVMDeposited {
 	validEvents := make([]*gatewayevm.GatewayEVMDeposited, 0)
-	for _, log := range rawLogs {
-		err := common.ValidateEvmTxLog(&log, gatewayAddr, "", common.TopicsGatewayDeposit)
+	for _, ethlog := range ethlogs {
+		err := common.ValidateEvmTxLog(&ethlog, gatewayAddr, "", common.TopicsGatewayDeposit)
 		if err != nil {
 			continue
 		}
-		depositedEvent, err := gatewayContract.ParseDeposited(log)
+		depositedEvent, err := gatewayContract.ParseDeposited(ethlog)
 		if err != nil {
 			ob.Logger().Inbound.Warn().
-				Stringer(logs.FieldTx, log.TxHash).
-				Uint64(logs.FieldBlock, log.BlockNumber).
+				Stringer(logs.FieldTx, ethlog.TxHash).
+				Uint64(logs.FieldBlock, ethlog.BlockNumber).
 				Msg("invalid Deposit event")
 			continue
 		}
@@ -219,7 +219,6 @@ func (ob *Observer) observeGatewayCall(
 		msg := ob.newCallInboundVote(event)
 
 		ob.Logger().Inbound.Info().
-			Str(logs.FieldMethod, "observeGatewayCall").
 			Stringer(logs.FieldTx, event.Raw.TxHash).
 			Uint64(logs.FieldBlock, event.Raw.BlockNumber).
 			Stringer("from", event.Sender).
@@ -237,12 +236,12 @@ func (ob *Observer) observeGatewayCall(
 
 // parseAndValidateCallEvents collects and sorts events by block number, tx index, and log index
 func (ob *Observer) parseAndValidateCallEvents(
-	rawLogs []ethtypes.Log,
+	ethlogs []ethtypes.Log,
 	gatewayAddr ethcommon.Address,
 	gatewayContract *gatewayevm.GatewayEVM,
 ) []*gatewayevm.GatewayEVMCalled {
 	validEvents := make([]*gatewayevm.GatewayEVMCalled, 0)
-	for _, log := range rawLogs {
+	for _, log := range ethlogs {
 		err := common.ValidateEvmTxLog(&log, gatewayAddr, "", common.TopicsGatewayCall)
 		if err != nil {
 			continue
@@ -328,7 +327,6 @@ func (ob *Observer) observeGatewayDepositAndCall(
 		msg := ob.newDepositAndCallInboundVote(event)
 
 		ob.Logger().Inbound.Info().
-			Str(logs.FieldMethod, "observeGatewayDepositAndCall").
 			Stringer(logs.FieldTx, event.Raw.TxHash).
 			Uint64(logs.FieldBlock, event.Raw.BlockNumber).
 			Stringer("from", event.Sender).